refactor: reorganise some utils classes
