+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/session/EmbraceMemoryCleanerService.kt
@@ -2,7 +2,6 @@ package io.embrace.android.embracesdk.internal.session
 
 import io.embrace.android.embracesdk.internal.logging.EmbLogger
 import io.embrace.android.embracesdk.internal.logging.InternalErrorType
-import io.embrace.android.embracesdk.internal.utils.stream
 import java.util.concurrent.CopyOnWriteArrayList
 
 internal class EmbraceMemoryCleanerService(private val logger: EmbLogger) : MemoryCleanerService {
@@ -14,7 +13,7 @@ internal class EmbraceMemoryCleanerService(private val logger: EmbLogger) : Memo
     val listeners: CopyOnWriteArrayList<MemoryCleanerListener> = CopyOnWriteArrayList<MemoryCleanerListener>()
 
     override fun cleanServicesCollections() {
-        stream(listeners) { listener: MemoryCleanerListener ->
+        listeners.toList().forEach { listener: MemoryCleanerListener ->
             try {
                 listener.cleanCollections()
             } catch (ex: Exception) {
