--- embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/utils/MainThreadUtils.kt
@@ -1,24 +1,0 @@
-package io.embrace.android.embracesdk.internal.utils
-
-import android.os.Handler
-import android.os.Looper
-
-object MainThreadUtils {
-
-    private val mainLooper = Looper.getMainLooper()
-    private val mainThread = mainLooper.thread
-
-    fun runOnMainThread(runnable: Runnable) {
-        val wrappedRunnable = Runnable {
-            runCatching {
-                runnable.run()
-            }
-        }
-        if (Thread.currentThread() !== mainThread) {
-            val mainHandler = Handler(mainLooper)
-            mainHandler.post(wrappedRunnable)
-        } else {
-            wrappedRunnable.run()
-        }
-    }
-}
