+++ embrace-android-instrumentation-huc/src/main/java/io/embrace/android/embracesdk/instrumentation/huc/EmbraceUrlConnectionDelegate.java
@@ -33,7 +33,6 @@ import javax.net.ssl.SSLSocketFactory;
 import io.embrace.android.embracesdk.annotation.InternalApi;
 import io.embrace.android.embracesdk.internal.instrumentation.network.EmbraceHttpPathOverrideKt;
 import io.embrace.android.embracesdk.internal.network.http.NetworkCaptureData;
-import io.embrace.android.embracesdk.internal.utils.exceptions.function.CheckedSupplier;
 import io.embrace.android.embracesdk.network.EmbraceNetworkRequest;
 import io.embrace.android.embracesdk.network.http.HttpMethod;
 import kotlin.jvm.functions.Function0;
